feat: multi mode
